SiVista_BE/requirements.txt
adrf==0.1.6
altgraph==0.17.4
amqp==5.2.0
asgiref==3.8.1
astroid==3.3.9
async-property==0.2.2
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.34.144
botocore==1.34.144
cachetools==5.5.2
celery==5.4.0
certifi==2024.6.2
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
daphne==4.1.2
dill==0.3.9
Django==5.1.3
django-cors-headers==4.3.1
django-mongodb-engine==0.6.0
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djangotoolbox==1.8.0
dnspython==2.7.0
drf-yasg==1.21.7
et_xmlfile==2.0.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.65.4
grpcio-tools==1.65.4
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==2.1.5
mccabe==0.7.0
msgpack==1.1.0
numpy==2.1.0
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==1.7.1
pylint==3.3.5
pymongo==4.10.1
pyOpenSSL==24.2.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.1.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9
s3transfer==0.10.2
service-identity==24.1.0
setuptools==72.2.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.3.1
tomlkit==0.13.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3
elasticsearch==8.12.0
